+++ crawlab/flower.py
@@ -3,14 +3,18 @@ import sys
 import subprocess
 
 # make sure the working directory is in system path
-file_dir = os.path.dirname(os.path.realpath(__file__))
-root_path = os.path.abspath(os.path.join(file_dir, '..'))
-sys.path.append(root_path)
+FILE_DIR = os.path.dirname(os.path.realpath(__file__))
+ROOT_PATH = os.path.abspath(os.path.join(FILE_DIR, '..'))
+sys.path.append(ROOT_PATH)
+
 from utils.log import other
 from config import BROKER_URL
 
 if __name__ == '__main__':
-    p = subprocess.Popen(['celery', 'flower', '-b', BROKER_URL], stdout=subprocess.PIPE, stderr=subprocess.STDOUT)
+    p = subprocess.Popen([sys.executable, '-m', 'celery', 'flower', '-b', BROKER_URL],
+                         stdout=subprocess.PIPE,
+                         stderr=subprocess.STDOUT,
+                         cwd=ROOT_PATH)
     for line in iter(p.stdout.readline, 'b'):
         if line.decode('utf-8') != '':
             other.info(line.decode('utf-8'))