+++ frontend/src/components/Config/ConfigList.vue
@@ -5,15 +5,27 @@
                :title="$t('Preview Results')"
                width="90%"
                :before-close="onDialogClose">
+      <el-table class="table-header" :data="[{}]" :show-header="false">
+        <el-table-column v-for="(f, index) in fields"
+                         :key="f.name + '-' + index"
+                         min-width="100px">
+          <template>
+            <el-input v-model="columnsDict[f.name]" size="mini" style="width: calc(100% - 15px)"></el-input>
+            <a href="javascript:" style="margin-left: 2px;" @click="onDeleteField(index)">X</a>
+            <!--<el-button size="mini" type="danger" icon="el-icon-delete" style="width:45px;margin-left:2px"></el-button>-->
+          </template>
+        </el-table-column>
+      </el-table>
       <el-table :data="previewCrawlData"
-                :header-cell-style="{background:'rgb(48, 65, 86)',color:'white'}"
+                :show-header="false"
                 border>
         <el-table-column v-for="(f, index) in fields"
+                         :key="f.name + '-' + index"
                          :label="f.name"
-                         :key="index"
                          min-width="100px">
+
           <template slot-scope="scope">
-            {{scope.row[f.name]}}
+            {{getDisplayStr(scope.row[f.name])}}
           </template>
         </el-table-column>
       </el-table>
@@ -22,8 +34,8 @@
 
     <!--config detail-->
     <el-row>
-      <el-col :span="11" :offset="1">
-        <el-form label-width="150px">
+      <el-form label-width="150px" ref="form" :model="spiderForm">
+        <el-col :span="11" :offset="1">
           <el-form-item :label="$t('Crawl Type')">
             <el-button-group>
               <el-button v-for="type in crawlTypeList"
@@ -34,7 +46,7 @@
               </el-button>
             </el-button-group>
           </el-form-item>
-          <el-form-item :label="$t('Start URL')" required>
+          <el-form-item :label="$t('Start URL')" prop="start_url" required>
             <el-input v-model="spiderForm.start_url" :placeholder="$t('Start URL')"></el-input>
           </el-form-item>
           <el-form-item :label="$t('Obey robots.txt')">
@@ -43,10 +55,8 @@
           <!--<el-form-item :label="$t('URL Pattern')">-->
           <!--<el-input v-model="spiderForm.url_pattern" :placeholder="$t('URL Pattern')"></el-input>-->
           <!--</el-form-item>-->
-        </el-form>
-      </el-col>
-      <el-col :span="11" :offset="1">
-        <el-form label-width="150px">
+        </el-col>
+        <el-col :span="11" :offset="1">
           <el-form-item :label="$t('Item Selector')"
                         v-if="['list','list-detail'].includes(spiderForm.crawl_type)">
             <el-select style="width: 35%;margin-right: 10px;"
@@ -73,8 +83,12 @@
                       :placeholder="$t('Pagination Selector')">
             </el-input>
           </el-form-item>
-        </el-form>
-      </el-col>
+          <el-form-item :label="$t('Item Threshold')"
+                        v-if="['list','list-detail'].includes(spiderForm.crawl_type)">
+            <el-input-number v-model="spiderForm.item_threshold"/>
+          </el-form-item>
+        </el-col>
+      </el-form>
     </el-row>
     <!--./config detail-->
 
@@ -134,7 +148,8 @@ export default {
       extractFieldsLoading: false,
       previewLoading: false,
       saveLoading: false,
-      dialogVisible: false
+      dialogVisible: false,
+      columnsDict: {}
     }
   },
   computed: {
@@ -189,23 +204,33 @@ export default {
     },
     onDialogClose () {
       this.dialogVisible = false
+      this.fields.forEach(f => {
+        f.name = this.columnsDict[f.name]
+      })
     },
     onPreview () {
-      this.onSave()
-        .then(() => {
-          this.previewLoading = true
-          this.$store.dispatch('spider/getPreviewCrawlData')
+      this.$refs['form'].validate(res => {
+        if (res) {
+          this.onSave()
             .then(() => {
-              this.dialogVisible = true
+              this.previewLoading = true
+              this.$store.dispatch('spider/getPreviewCrawlData')
+                .then(() => {
+                  this.fields.forEach(f => {
+                    this.columnsDict[f.name] = f.name
+                  })
+                  this.dialogVisible = true
+                })
+                .catch(() => {
+                  this.$message.error(this.$t('Something wrong happened'))
+                })
+                .finally(() => {
+                  this.previewLoading = false
+                })
+              this.$st.sendEv('爬虫详情-配置', '预览')
             })
-            .catch(() => {
-              this.$message.error(this.$t('Something wrong happened'))
-            })
-            .finally(() => {
-              this.previewLoading = false
-            })
-          this.$st.sendEv('爬虫详情-配置', '预览')
-        })
+        }
+      })
     },
     onCrawl () {
       this.$confirm(this.$t('Are you sure to run this spider?'), this.$t('Notification'), {
@@ -221,25 +246,42 @@ export default {
         })
     },
     onExtractFields () {
-      this.onSave()
-        .then(() => {
-          this.extractFieldsLoading = true
-          this.$store.dispatch('spider/extractFields')
-            .then(response => {
-              if (response.data.item_selector) {
-                this.$set(this.spiderForm, 'item_selector', response.data.item_selector)
-                this.$set(this.spiderForm, 'item_selector_type', 'css')
-              }
+      this.$refs['form'].validate(res => {
+        if (res) {
+          this.onSave()
+            .then(() => {
+              this.extractFieldsLoading = true
+              this.$store.dispatch('spider/extractFields')
+                .then(response => {
+                  if (response.data.item_selector) {
+                    this.$set(this.spiderForm, 'item_selector', response.data.item_selector)
+                    this.$set(this.spiderForm, 'item_selector_type', 'css')
+                  }
 
-              if (response.data.fields && response.data.fields.length) {
-                this.spiderForm.fields = response.data.fields
-              }
+                  if (response.data.fields && response.data.fields.length) {
+                    this.spiderForm.fields = response.data.fields
+                  }
+
+                  if (response.data.pagination_selector) {
+                    this.spiderForm.pagination_selector = response.data.pagination_selector
+                  }
+                })
+                .finally(() => {
+                  this.extractFieldsLoading = false
+                })
+              this.$st.sendEv('爬虫详情-配置', '提取字段')
             })
-            .finally(() => {
-              this.extractFieldsLoading = false
-            })
-          this.$st.sendEv('爬虫详情-配置', '提取字段')
-        })
+        }
+      })
+    },
+    onDeleteField (index) {
+      this.fields.splice(index, 1)
+    },
+    getDisplayStr (value) {
+      if (!value) return value
+      value = value.trim()
+      if (value.length > 20) return value.substr(0, 20) + '...'
+      return value
     }
   },
   created () {
@@ -248,7 +290,7 @@ export default {
       this.spiderForm.fields = []
       for (let i = 0; i < 3; i++) {
         this.spiderForm.fields.push({
-          name: `field_${i + 1}`,
+          name: 'field_' + (i + 1),
           type: 'css',
           extract_type: 'text'
         })
@@ -260,7 +302,7 @@ export default {
       this.spiderForm.detail_fields = []
       for (let i = 0; i < 3; i++) {
         this.spiderForm.detail_fields.push({
-          name: `field_${i + 1}`,
+          name: 'field_' + (i + 1),
           type: 'css',
           extract_type: 'text'
         })
@@ -268,10 +310,11 @@ export default {
     }
 
     if (!this.spiderForm.crawl_type) this.$set(this.spiderForm, 'crawl_type', 'list')
-    if (!this.spiderForm.start_url) this.$set(this.spiderForm, 'start_url', 'http://example.com')
+    // if (!this.spiderForm.start_url) this.$set(this.spiderForm, 'start_url', 'http://example.com')
     if (!this.spiderForm.item_selector_type) this.$set(this.spiderForm, 'item_selector_type', 'css')
     if (!this.spiderForm.pagination_selector_type) this.$set(this.spiderForm, 'pagination_selector_type', 'css')
-    if (this.spiderForm.obey_robots_txt === undefined) this.$set(this.spiderForm, 'obey_robots_txt', true)
+    if (this.spiderForm.obey_robots_txt == null) this.$set(this.spiderForm, 'obey_robots_txt', true)
+    if (this.spiderForm.item_threshold == null) this.$set(this.spiderForm, 'item_threshold', 10)
   }
 }
 </script>
@@ -301,5 +344,17 @@ export default {
   .title {
     color: #606266;
     font-size: 14px;
+  }
+
+  .el-table.table-header >>> td {
+    padding: 0;
+  }
+
+  .el-table.table-header >>> .cell {
+    padding: 0;
+  }
+
+  .el-table.table-header >>> .el-input .el-input__inner {
+    border-radius: 0;
   }
 </style>